Move batch policy to batch folder
